+++ src/main/java/org/apache/aurora/scheduler/storage/db/typehandlers/TypeHandlers.java
@@ -35,6 +35,7 @@ public final class TypeHandlers {
         .add(MaintenanceModeTypeHandler.class)
         .add(ScheduleStatusTypeHandler.class)
         .add(ResourceTypeHandler.class)
+        .add(VolumeModeTypeHandler.class)
         .build();
   }
 }