+++ src/test/java/org/apache/aurora/scheduler/storage/AbstractTaskStoreTest.java
@@ -51,8 +51,10 @@ import org.apache.aurora.gen.MaintenanceMode;
 import org.apache.aurora.gen.MesosContainer;
 import org.apache.aurora.gen.MesosFetcherURI;
 import org.apache.aurora.gen.Metadata;
+import org.apache.aurora.gen.Mode;
 import org.apache.aurora.gen.ScheduledTask;
 import org.apache.aurora.gen.TaskQuery;
+import org.apache.aurora.gen.Volume;
 import org.apache.aurora.scheduler.base.JobKeys;
 import org.apache.aurora.scheduler.base.Query;
 import org.apache.aurora.scheduler.base.TaskTestUtil;
@@ -187,6 +189,20 @@ public abstract class AbstractTaskStoreTest extends TearDownTestCase {
             new MesosFetcherURI("pathA").setExtract(true).setCache(true),
             new MesosFetcherURI("pathB").setExtract(true).setCache(true)));
     IScheduledTask task = IScheduledTask.build(builder);
+    saveTasks(task);
+    assertStoreContents(task);
+  }
+
+  @Test
+  public void testSaveWithContainerVolumes() {
+    ScheduledTask builder = TASK_B.newBuilder();
+    Image image = Image.docker(new DockerImage().setName("some-name").setTag("some-tag"));
+    List<Volume> volumes = ImmutableList.of(new Volume("container", "host", Mode.RO));
+    builder.getAssignedTask().getTask().getContainer().getMesos().setImage(image)
+        .setVolumes(volumes);
+
+    IScheduledTask task = IScheduledTask.build(builder);
+
     saveTasks(task);
     assertStoreContents(task);
   }